requirements.txt
opencv-python<4.9
face_recognition
numpy<2
python-dotenv
reportlab
tkcalendar
Pillow
bcrypt